pyproject.toml
[project]
name = "sonarqube-analysis-agent"
version = "1.0.0"
description = "AI agent for analyzing SonarQube findings and generating automated fixes"
authors = [
    {name = "SonarQube Agent Team"}
]
requires-python = ">=3.10"
dependencies = [
    "llama-stack>=0.0.13",
    "llama-stack-client>=0.0.13",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "asyncio>=3.4.3",
    "pydantic>=2.5.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
